+++ cli/cmd/migrate.go
@@ -62,7 +62,7 @@ func migrate(cmd *cobra.Command, args []string) error {
 		transformerSpecsByName[transformer.Name] = *transformer
 	}
 
-	authToken, err := auth.GetAuthTokenIfNeeded(log.Logger, sources, destinations)
+	authToken, err := auth.GetAuthTokenIfNeeded(log.Logger, sources, destinations, transformers)
 	if err != nil {
 		return fmt.Errorf("failed to get auth token: %w", err)
 	}