+++ plugins/destination/kafka/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [5.1.2](https://github.com/cloudquery/cloudquery/compare/plugins-destination-kafka-v5.1.1...plugins-destination-kafka-v5.1.2) (2024-09-18)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.62.0 ([#19137](https://github.com/cloudquery/cloudquery/issues/19137)) ([ed315d0](https://github.com/cloudquery/cloudquery/commit/ed315d011d6a205e1a8ba851570f8e9533698c52))
+
 ## [5.1.1](https://github.com/cloudquery/cloudquery/compare/plugins-destination-kafka-v5.1.0...plugins-destination-kafka-v5.1.1) (2024-09-04)